app/requirements.txt
fastapi
pydantic
uvicorn
pymongo
pandas
python-dotenv


